rename and move things
