+++ api/src/modules/Session/model.ts
@@ -1,13 +1,10 @@
-import type { PrismaClient, Prisma } from '@prisma/client';
+import type { PrismaClient } from '@prisma/client';
 import { inputObjectType, objectType } from 'nexus/dist';
-import { randomUUID } from 'crypto';
 import { GraphQLError } from 'graphql';
 import ms from 'ms';
-import wildcard from 'wildcard-match';
 import { token as tokenUtils } from '../../helpers';
 import { JWT_EXPIRATION_PERIOD, OWN_ORIGIN } from '../../env';
-
-// Models
+import { validateOriginAgainstAllowed, generateTokenAndSession } from './helpers';
 
 export const Session = objectType({
   name: 'Session',
@@ -40,86 +37,6 @@ export const SessionCreateOutput = objectType({
     t.nonNull.string('token');
   },
 });
-
-// Actions
-
-function parseOriginMarkup(originParam: string): string {
-  if (originParam === '*') {
-    return '*';
-  }
-  const trimmedOriginParam = originParam.trim();
-  const origins = trimmedOriginParam.split(',').map((origin) => origin.trim());
-  origins.forEach((origin) => {
-    if (!origin.startsWith('http://') && !origin.startsWith('https://')) {
-      throw new Error("Origin must start with 'http://' or 'https://'");
-    }
-  });
-  return origins.join(',');
-}
-
-function throwGQLErrorIfOriginDisallowed(
-  allowedOrigins: string,
-  originReceived?: string,
-) {
-  if (allowedOrigins === '*') {
-    return;
-  }
-  if (!originReceived) {
-    throw new GraphQLError('Origin not provided', {
-      extensions: { code: 'ORIGIN_HEADER_MISSING' },
-    });
-  }
-  const allowedOriginsSplit = allowedOrigins.split(',');
-  if (!wildcard(allowedOriginsSplit)(originReceived)) {
-    throw new GraphQLError('Access denied due to origin restriction', {
-      extensions: { code: 'ORIGIN_FORBIDDEN' },
-    });
-  }
-}
-async function newSession(
-  prisma: PrismaClient,
-  session: Prisma.SessionCreateInput,
-) {
-  return prisma.session.create({
-    data: session,
-  });
-}
-
-const generateTokenAndSession = async (
-  prisma: PrismaClient,
-  userId: string,
-  session: { expiryDurationSeconds?: number | null; name: string; allowedOrigins: string },
-  isUserCreated: boolean = false,
-) => {
-  const createId = randomUUID();
-  const createdToken = tokenUtils.generate(createId, session.expiryDurationSeconds);
-  const expiryDate = tokenUtils.getExpiryDateFromToken(createdToken);
-  const formattedToken = tokenUtils.format(createdToken);
-
-  let parsedAllowedOrigins: string = '';
-  try { parsedAllowedOrigins = parseOriginMarkup(session.allowedOrigins); } catch (e: any) {
-    throw new GraphQLError(e.message, { extensions: { code: 'INVALID_ORIGIN_FORMAT' } });
-  }
-
-  const createData = {
-    allowedOrigins: parsedAllowedOrigins,
-    name: session.name,
-    referenceExpiryDate: expiryDate,
-    id: createId,
-    referenceTokenId: formattedToken,
-    isUserCreated,
-    creator: {
-      connect: {
-        id: userId,
-      },
-    },
-  };
-  const createdSession = await newSession(prisma, createData);
-  return {
-    token: createdToken,
-    session: createdSession,
-  };
-};
 
 export function getSessionCrud(prisma: PrismaClient) {
   return {
@@ -198,7 +115,7 @@ export function getSessionCrud(prisma: PrismaClient) {
           extensions: { code: 'SESSION_EXPIRED' },
         });
       }
-      throwGQLErrorIfOriginDisallowed(session.allowedOrigins, origin);
+      validateOriginAgainstAllowed(session.allowedOrigins, origin);
       return session;
     },
